fix: 스켈레톤 수정
