+++ src/components/common/Skeleton/Skeleton.tsx
@@ -1,4 +1,5 @@
 import cn from "@/lib/utils";
+import { PropsWithChildren } from "react";
 
 function Skeleton({ className, ...props }: React.ComponentProps<"div">) {
   return (
@@ -9,5 +10,51 @@ function Skeleton({ className, ...props }: React.ComponentProps<"div">) {
     />
   );
 }
+
+function SkeletonWrapper({ children }: PropsWithChildren) {
+  return <div className="flex flex-col gap-2.5">{children}</div>;
+}
+
+function SkeletonInput({ className, ...props }: React.ComponentProps<"div">) {
+  return (
+    <Skeleton
+      className={cn(
+        "h-9 w-full rounded-lg md:w-full lg:h-12 lg:rounded-xl",
+        className
+      )}
+      {...props}
+    />
+  );
+}
+
+function SkeletonLabel({ className, ...props }: React.ComponentProps<"div">) {
+  return (
+    <Skeleton
+      className={cn("h-[12px] w-20 lg:h-[13px]", className)}
+      {...props}
+    />
+  );
+}
+
+function SkeletonFieldGroup({ total }: { total: number }) {
+  return (
+    <>
+      {Array.from({ length: total })
+        .fill(0)
+        .map((_, i) => (
+          // eslint-disable-next-line react/no-array-index-key
+          <div className="flex flex-col gap-2.5" key={i}>
+            <SkeletonLabel />
+            <SkeletonInput />
+          </div>
+        ))}
+    </>
+  );
+}
+
+Skeleton.Wrapper = SkeletonWrapper;
+Skeleton.Input = SkeletonInput;
+Skeleton.Label = SkeletonLabel;
+Skeleton.FieldGroup = SkeletonFieldGroup;
 
 export { Skeleton };
