--- src/components/common/Skeleton/SkeletonInput.tsx
@@ -1,7 +1,0 @@
-import { Skeleton } from "./Skeleton";
-
-export default function SkeletonInput() {
-  return (
-    <Skeleton className="h-9 w-full rounded-lg md:w-full lg:h-12 lg:rounded-xl" />
-  );
-}
