--- src/components/common/Skeleton/SkeletonLabel.tsx
@@ -1,5 +1,0 @@
-import { Skeleton } from "./Skeleton";
-
-export default function SkeletonLabel() {
-  return <Skeleton className="h-[12px] w-20 lg:h-[13px]" />;
-}
